// src/usuario/dto/create-usuario.dto.ts
export class CreateUsuarioDto {
    ID: number;
    NOMBRES: string;
    APELLIDOS: string;
    DOCUMENTO: String;
    CLAVEINGRESO: string;
    FECHANACIMIENTO: string;
    NUMEROTELEFONO: number;
    NUMEROMOVIL: number;
    CORREO: string;
    DIRECCION: string;
    ESTADO: number;
    CODIGOTIPOUSUARIO: number;
}